engine: update tests
